+++ package.json
@@ -27,6 +27,7 @@
     "chartjs-plugin-datalabels": "0.7.0",
     "flatpickr": "4.6.3",
     "he": "1.2.0",
+    "lodash": "4.17.15",
     "moment": "2.24.0",
     "nanoid": "2.1.9"
   }
