+++ public/css/style.css
@@ -267,7 +267,10 @@ main {
   opacity: 1;
   cursor: default; }
 
-.trip-search {
+.trip-filters__filter-input:disabled + .trip-filters__filter-label {
+  cursor: not-allowed; }
+
+  .trip-search {
   -ms-flex-positive: 1;
       flex-grow: 1;
   display: -ms-flexbox;
