+++ src/components/abstract-smart-component.js
@@ -1,10 +1,6 @@
 import AbstractComponent from './abstract-component.js';
 
 export default class AbstractSmartComponent extends AbstractComponent {
-  recoveryListeners() {
-    throw new Error(`Abstract method not implemented: recoveryListeners`);
-  }
-
   rerender() {
     const oldElement = this.getElement();
     const parent = oldElement.parentElement;
@@ -13,7 +9,5 @@ export default class AbstractSmartComponent extends AbstractComponent {
     const newElement = this.getElement();
 
     parent.replaceChild(newElement, oldElement);
-
-    this.recoveryListeners();
   }
 }
