+++ src/components/trip-info.js
@@ -44,8 +44,6 @@ export default class TripInfo extends AbstractSmartComponent {
      `;
   }
 
-  recoveryListeners() {}
-
   rerender(events) {
     this._events = events.slice().sort((a, b) => a.startDate - b.startDate);
