+++ src/main.js
@@ -2,26 +2,26 @@ import Api from './api/api.js';
 import Store from './api/store.js';
 import Provider from './api/provider.js';
 import TripController from './controllers/trip-controller.js';
-import StatisticsComponent from './components/statistics.js';
 import FilterController from './controllers/filter-controller.js';
-import EventsModel from './models/events-model.js';
-import LoadEvents from './components/load-events.js';
+import StatisticsController from './controllers/statistics-controller';
 import SiteMenuComponent from './components/menu.js';
+import LoadEvents from './components/load-events.js';
+import EventsModel from './models/events-model.js';
+import EventModel from './models/event-model.js';
 import {renderElement, removeElement, RenderPosition} from './utils/render.js';
-import {MenuItem, AUTHORIZATION, END_POINT} from './const.js';
+import {MenuItem, AUTHORIZATION, END_POINT, STORE_NAME} from './const.js';
 
-const STORE_PREFIX = `bigtrip-localstorage`;
-const STORE_VER = `v1`;
-const STORE_NAME = `${STORE_PREFIX}-${STORE_VER}`;
 
 const tripControlsElement = document.querySelector(`.trip-main__trip-controls`);
 const tripEventsElement = document.querySelector(`.trip-events`);
 const pageMainElement = document.querySelector(`.page-main`);
+const bodyContainerElement = pageMainElement.querySelector(`.page-body__container`);
+const eventAddButtonElement = document.querySelector(`.trip-main__event-add-btn`);
 
 window.addEventListener(`load`, () => {
   navigator.serviceWorker.register(`/sw.js`)
-    .then(() => {
-    }).catch((err) => {
+    .then(() => {})
+    .catch((err) => {
       throw new Error(err);
     });
 });
@@ -43,31 +43,27 @@ const menuItems = Object.values(MenuItem)
 const menuComponent = new SiteMenuComponent(menuItems);
 const filterController = new FilterController(tripControlsElement, eventsModel);
 const tripController = new TripController(tripEventsElement, eventsModel, apiWithProvider);
+const statisticsController = new StatisticsController(bodyContainerElement, eventsModel);
 const loadEvents = new LoadEvents();
-const statisticsComponent = new StatisticsComponent(eventsModel);
 
 renderElement(tripControlsElement, menuComponent, RenderPosition.AFTERBEGIN);
 filterController.render();
 renderElement(tripEventsElement, loadEvents);
-renderElement(pageMainElement.querySelector(`.page-body__container`), statisticsComponent);
 
-statisticsComponent.hide();
-
-document.querySelector(`.trip-main__event-add-btn`)
-   .addEventListener(`click`, () => {
-     tripController.createEvent();
-   });
+eventAddButtonElement.addEventListener(`click`, () => {
+  tripController.createEvent();
+});
 
 menuComponent.setItemClickHandler((menuItem) => {
   switch (menuItem) {
     case MenuItem.STATS:
       menuComponent.setActiveItem(MenuItem.STATS);
       tripController.hide();
-      statisticsComponent.show();
+      statisticsController.show();
       break;
     case MenuItem.TABLE:
       menuComponent.setActiveItem(MenuItem.TABLE);
-      statisticsComponent.hide();
+      statisticsController.hide();
       tripController.show();
       break;
   }
@@ -82,6 +78,8 @@ Promise.all([apiWithProvider.getDestinations(), apiWithProvider.getOffers(), api
      eventsModel.setEvents(events);
      tripController.render();
      removeElement(loadEvents);
+     statisticsController.render();
+     statisticsController.hide();
    });
 
 window.addEventListener(`online`, () => {
@@ -89,12 +87,13 @@ window.addEventListener(`online`, () => {
 
   if (!apiWithProvider.getSynchronize()) {
     apiWithProvider.sync()
-        .then(() => {
-          // Действие, в случае успешной синхронизации
-        })
-        .catch((err) => {
-          throw new Error(err);
-        });
+    .then((events) => {
+      eventsModel.setEvents(EventModel.parseEvents(events));
+      tripController.updateEvents();
+    })
+    .catch((err) => {
+      throw new Error(err);
+    });
   }
 });
